chore: created new components for small map and details sub header
